+++ FFCalendar/FFCalendars/Util/FFEventDetail and FFEditEvent Views and Popovers/FFEditEvent View and Popover/FFSearchBarWithAutoComplete/FFSearchBarWithAutoComplete.m
@@ -87,7 +87,7 @@
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     
-    int intRows = [arrayOfTableView count];
+    long intRows = [arrayOfTableView count];
     
     CGFloat height = 4*BUTTON_HEIGHT;
     if (intRows < 4) {
